[NEW RBO] Combined push rename into producer rule
