+++ ydb/core/kqp/opt/rbo/rules/ya.make
@@ -32,11 +32,8 @@ SRCS(
     push_olap_projection.cpp
     disable_blocks_on_columns_limit.cpp
     map/rename_to_append.cpp
-    map/rename_common.cpp
     map/remove_identity_map.cpp
-    map/push_rename_into_aggregate_result.cpp
-    map/push_rename_into_map_producer.cpp
-    map/push_rename_into_read.cpp
+    map/push_rename_into_producer.cpp
     map/rewrite_to_preferred_alias.cpp
     propagate_aggregate_through_stage.cpp
     propagate_hash_func_stage.cpp
